requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
babel==2.17.0
beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
courlan==1.3.2
cymem==2.0.11
dateparser==1.2.2
dnspython==2.7.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
feedparser==6.0.11
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.8
htmldate==1.9.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jusText==3.0.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
ollama==0.5.3
packaging==25.0
pandas==2.3.1
pillow==11.3.0
preshed==3.0.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
safetensors==0.6.2
sentry-sdk==2.35.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.2
sympy==1.14.0
thinc==8.3.6
tld==0.13.1
tldextract==5.3.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.55.4
triton==3.4.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
